Moved constraints and ecf_api to appropriate locations
