+++ app/models/external/ecf_api/base.rb
