+++ app/models/external/ecf_api/npq/base.rb
