+++ app/models/external/ecf_api/npq/user.rb
