+++ app/models/external/ecf_api/npq_course.rb
