+++ app/models/external/ecf_api/user.rb
