+++ config/routes.rb
@@ -61,7 +61,7 @@ Rails.application.routes.draw do
       resources :processing_jobs, only: %i[create], controller: "webhook_messages/processing_jobs"
     end
 
-    constraints RouteConstraints::HasFlipperAccess do
+    constraints HasFlipperAccess do
       mount Flipper::UI.app(Flipper) => "/feature_flags"
     end
   end
